pyproject.toml
[tool.poetry]
name = "hh-helper"
version = "0.1.0"
description = "HH.ru helper for everyday tasks"
authors = ["plat01"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
aiohttp = "^3.11.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
